requirements.txt
Flask==2.2.1
google-cloud-secret-manager
